+++ .gitignore
@@ -1,0 +1,3 @@
+.idea/
+build
+opensearch-hebrew-analyser/.gradle
