--- Hebrew-ElasticSearch-ngrams-3-words/es-plugin.properties
@@ -1,1 +1,0 @@
-plugin=com.hotstar.hebrew.plugin.HebrewAnalyzerEsPlugin
