--- Hebrew-ElasticSearch-ngrams-3-words/src/main/java/com/hotstar/hebrew/plugin/HebrewNoOpTokenFilterFactory.java
@@ -1,19 +1,0 @@
-package com.hotstar.hebrew.plugin;
-
-
-import org.apache.lucene.analysis.TokenStream;
-import org.elasticsearch.common.settings.Settings;
-import org.elasticsearch.env.Environment;
-import org.elasticsearch.index.IndexSettings;
-import org.elasticsearch.index.analysis.AbstractTokenFilterFactory;
-
-public class HebrewNoOpTokenFilterFactory extends AbstractTokenFilterFactory {
-    public HebrewNoOpTokenFilterFactory(IndexSettings indexSettings, Environment environment, String name, Settings settings) {
-        super(name, settings);
-    }
-
-    @Override
-    public TokenStream create(TokenStream tokenStream) {
-        return tokenStream;
-    }
-}
