--- Hebrew-ElasticSearch-ngrams-3-words/src/main/java/com/hotstar/hebrew/plugin/HebrewTokenizerTokenizerFactory.java
@@ -1,20 +1,0 @@
-package com.hotstar.hebrew.plugin;
-
-import org.apache.lucene.analysis.Tokenizer;
-import org.elasticsearch.common.settings.Settings;
-import org.elasticsearch.env.Environment;
-import org.elasticsearch.index.IndexSettings;
-import org.elasticsearch.index.analysis.AbstractTokenizerFactory;
-import com.hotstar.hebrew.analysis.SefariaTokenizer;
-
-
-public class HebrewTokenizerTokenizerFactory extends AbstractTokenizerFactory {
-    public HebrewTokenizerTokenizerFactory(IndexSettings indexSettings, Environment environment, String name, Settings settings) {
-        super(indexSettings, settings, name);
-    }
-
-    @Override
-    public Tokenizer create() {
-        return new SefariaTokenizer();
-    }
-}
