--- Hebrew-ElasticSearch-semi-exact/plugin-descriptor.properties
@@ -1,7 +1,0 @@
-jvm=true
-name=elasticsearch-analysis-hebrew
-description=elasticsearch-analysis-hebrew
-classname=com.hotstar.hebrew.plugin.HebrewAnalyzerEsPlugin
-elasticsearch.version=8.5.3
-java.version=17
-version=1.0
