--- Hebrew-ElasticSearch-semi-exact/src/main/java/com/hotstar/hebrew/analysis/HebrewTokenTypeAttribute.java
@@ -1,17 +1,0 @@
-package com.hotstar.hebrew.analysis;
-import org.apache.lucene.util.Attribute;
-
-/**
- * This attribute is used to pass info on tokens as parsed and identified
- * by the HebMorph tokenizer
- */
-public interface HebrewTokenTypeAttribute extends Attribute{
-    enum HebrewType {
-        Unknown
-    }
-
-    void setType(HebrewType type);
-    HebrewType getType();
-    boolean isExact();
-    void setExact(boolean isExact);
-}
