--- Hebrew-ElasticSearch-semi-exact/src/main/java/com/hotstar/hebrew/plugin/HebrewAnalyzerEsPlugin.java
@@ -1,33 +1,0 @@
-package com.hotstar.hebrew.plugin;
-import org.elasticsearch.index.analysis.AnalyzerProvider;
-import org.elasticsearch.index.analysis.TokenFilterFactory;
-import org.elasticsearch.index.analysis.TokenizerFactory;
-import org.elasticsearch.plugins.AnalysisPlugin;
-import org.elasticsearch.plugins.Plugin;
-import org.elasticsearch.indices.analysis.AnalysisModule.AnalysisProvider;
-import org.apache.lucene.analysis.Analyzer;
-
-import java.util.HashMap;
-import java.util.Map;
-
-import static java.util.Collections.singletonMap;
-
-public class HebrewAnalyzerEsPlugin extends Plugin implements AnalysisPlugin {
-
-    @Override
-    public Map<String, AnalysisProvider<TokenFilterFactory>> getTokenFilters() {
-        Map<String, AnalysisProvider<TokenFilterFactory>> tokenFilters = new HashMap<>();
-        return tokenFilters;
-    }
-
-    @Override
-    public Map<String, AnalysisProvider<TokenizerFactory>> getTokenizers() {
-        Map<String, AnalysisProvider<TokenizerFactory>> extra = new HashMap<>();
-        return extra;
-    }
-
-    @Override
-    public Map<String, AnalysisProvider<AnalyzerProvider<? extends Analyzer>>> getAnalyzers() {
-        return singletonMap("hebrew_semi_exact_analyzer", HebrewAnalyzerProvider::new);
-    }
-}
