--- Hebrew-ElasticSearch-semi-exact/src/main/java/com/hotstar/hebrew/plugin/SefariaBinderProcessor.java
@@ -1,15 +1,0 @@
-package com.hotstar.hebrew.plugin;
-
-import org.elasticsearch.indices.analysis.AnalysisModule;;
-
-//public class SefariaBinderProcessor extends AnalysisModule.AnalysisBinderProcessor {
-//
-//    /* This is the only function that you need. It simply adds our PlusSignAnalyzerProvider class
-//     * to a list of bindings.
-//     */
-//    @Override
-//    public void processAnalyzers(AnalyzersBindings analyzersBindings) {
-//        analyzersBindings.processAnalyzer(SefariaAnalyzerProvider.NAME,
-//                SefariaAnalyzerProvider.class);
-//    }
-//}
