+++ opensearch-hebrew-analyser/LICENSE.txt
