+++ opensearch-hebrew-analyser/es-plugin.properties
@@ -1,0 +1,1 @@
+plugin=com.hotstar.hebrew.plugin.HebrewAnalyzerOsPlugin
