+++ opensearch-hebrew-analyser/settings.gradle
@@ -1,0 +1,18 @@
+/*
+ * This file was generated by the Gradle 'init' task.
+ *
+ * This project uses @Incubating APIs which are subject to change.
+ */
+
+pluginManagement {
+    repositories{
+        mavenLocal()
+        mavenCentral()
+        google()
+        maven{
+            url "https://plugins.gradle.org/m2/"
+        }
+        gradlePluginPortal()
+    }
+}
+rootProject.name = "opensearch-analysis-hebrew"
