+++ opensearch-hebrew-analyser/src/main/java/com/hotstar/hebrew/analysis/NGramizer.java
