+++ opensearch-hebrew-analyser/src/main/java/com/hotstar/hebrew/plugin/HebrewNoOpTokenFilterFactory.java
@@ -1,0 +1,19 @@
+package com.hotstar.hebrew.plugin;
+
+
+import org.apache.lucene.analysis.TokenStream;
+import org.opensearch.common.settings.Settings;
+import org.opensearch.env.Environment;
+import org.opensearch.index.IndexSettings;
+import org.opensearch.index.analysis.AbstractTokenFilterFactory;
+
+public class HebrewNoOpTokenFilterFactory extends AbstractTokenFilterFactory {
+    public HebrewNoOpTokenFilterFactory(IndexSettings indexSettings, Environment environment, String name, Settings settings) {
+        super(indexSettings, name, settings);
+    }
+
+    @Override
+    public TokenStream create(TokenStream tokenStream) {
+        return tokenStream;
+    }
+}
